# FileStore performance enhancements