--- a/src/maggma/stores/mongolike.py
+++ b/src/maggma/stores/mongolike.py
@@ -467,9 +467,7 @@ class MongoStore(Store):
         if not isinstance(docs, list):
             docs = [docs]
 
-        for d in docs:
-            d = jsanitize(d, allow_bson=True)
-
+        for d in map(lambda x: jsanitize(x, allow_bson=True), docs):
             # document-level validation is optional
             validates = True
             if self.validator: